refactor: secondaryIndex가 1일 때의 기록 필드 컴포넌트로 분리
